pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "claude-gateway"
version = "0.1.0"
description = "Telegram gateway for Claude Code CLI - control Claude remotely via Telegram"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
keywords = ["claude", "telegram", "bot", "ai", "cli", "gateway"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Chat",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]
dependencies = [
    "python-telegram-bot>=20.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "websockets>=12.0",
    "pydantic>=2.0",
]

[project.scripts]
claude-gateway = "claude_gateway.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/claude-gateway"
Repository = "https://github.com/yourusername/claude-gateway"
Issues = "https://github.com/yourusername/claude-gateway/issues"

[tool.hatch.build.targets.wheel]
packages = ["claude_gateway"]

[tool.hatch.build.targets.wheel.shared-data]
# Include the frontend dist if it exists
# "claude_gateway/frontend/dist" = "share/claude-gateway/frontend/dist"
